+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 - Reload gRPC SN config with SIGHUP (#3874)
 - Support for creation of containers with initial placement policy (#3830)
 - `neofs-lens meta remove` command (#3891)
+- `neofs-lens fstree remove` command (#3892)
 
 ### Fixed
 - Resending the header after chunks have already been sent in object service `Get` handler (#3833)